V0.1

+++ EmployeeManager.h
@@ -267,7 +267,7 @@ public:
             }
 
 //            *((long double*) averageBumpGrade) = sumOfGrades / numOfEmployees;
-            printf("AverageBumpGradeBetweenSalaryByGroup: %.1f\n", ((long double) (sumOfGrades / numOfEmployees)));
+            printf("AverageBumpGradeBetweenSalaryByGroup: %.1f\n", ((double) (sumOfGrades / numOfEmployees)));
 
         } catch (...){//only bad alloc
             return ALLOCATION_ERROR;
